+++ rollup.config.js
@@ -3,10 +3,11 @@ import { terser } from 'rollup-plugin-terser';
 
 export default [
   {
-    input: 'src/js/main.js',
+    input: 'src/js/index.js',
     output: [
       {
         file: 'src/js.min.js',
+        format: 'cjs',
       },
     ],
     external: [...Object.keys(pkg.dependencies || {})],