+++ src/js/main.js
@@ -1,8 +1,0 @@
-import { updatePomodoroCounterDots, update } from './timer';
-import { applyActions } from './windowButtons';
-
-window.addEventListener('DOMContentLoaded', () => {
-  updatePomodoroCounterDots();
-  applyActions();
-  setInterval(update, 1000);
-});